+++ src/js/debug.js
@@ -2,6 +2,8 @@
     Debug functions
 */
 
+const globals = nodeRequire('./js/globals');
+
 module.exports = () => {
 
 };